use lazy import for joblib
It seems you have some configuration issues with `sklearn`. if you don't want to fix `sklearn` version in the virtual env then you should consider using lazy import. In the previous implementation, it tries anyway to import both and in SEPAL fail to import `sklearn.joblib`.
Also note that lazy import is a very bad practice in itself (banned in both Black formatter and Flake8), so if you consider distributing the lib to more people, maybe correcting the environment would be a good thing.